Move interest-related code to separate class
